backend/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-dotenv
pydantic
aiofiles
python-multipart
pydantic-settings
google-adk
google-adk-tools
pydub